fix: 修复Windows编译错误 - na::Matrix3和u64类型不匹配
- panorama_stitching.rs:511: na::Matrix3 → Matrix3 (使用已import的nalgebra::Matrix3)
- super_resolution.rs:182,185: .min(u64::MAX as usize) → .min(u64::MAX)
  (修复Windows上usize与u64类型严格不匹配错误)

这两个错误导致v1.7.5构建全平台失败。

+++ src-tauri/src/super_resolution.rs
@@ -179,10 +179,10 @@ pub fn apply_super_resolution(
     let out_h = height * scale;
 
     // Accumulator for the output image (float for blending)
-    let acc_len = (out_w as u64 * out_h as u64 * 4).min(u64::MAX as usize) as usize;
+    let acc_len = (out_w as u64 * out_h as u64 * 4).min(u64::MAX) as usize;
     let mut accumulator = vec![0.0f32; acc_len];
     // Weight accumulator for blending
-    let weight_len = (out_w as u64 * out_h as u64).min(u64::MAX as usize) as usize;
+    let weight_len = (out_w as u64 * out_h as u64).min(u64::MAX) as usize;
     let mut weight_accum = vec![0.0f32; weight_len];
 
     let step = tile_params.ucs.saturating_sub(tile_params.overlap).max(1);
